// source/web/src/api/generated/model/reportDetailsDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * AdventureWorksAIWorkspaceAPI.Api
 * OpenAPI spec version: 1.0
 */
import type { ReportDetailsDtoStatus } from './reportDetailsDtoStatus'
import type { ReportDetailsDtoResult } from './reportDetailsDtoResult'
import type { ChartSpec } from './chartSpec'
import type { ReportSectionDto } from './reportSectionDto'
import type { ReportMessageDto } from './reportMessageDto'
import type { GeneratedSqlQueryDto } from './generatedSqlQueryDto'

export interface ReportDetailsDto {
  id: string
  title: string
  originalPrompt: string
  /** @nullable */
  summary?: string | null
  /** @nullable */
  conclusions?: string | null
  status: ReportDetailsDtoStatus
  isFavorite: boolean
  createdAt: string
  updatedAt: string
  /** @nullable */
  result?: ReportDetailsDtoResult
  charts: ChartSpec[]
  sections: ReportSectionDto[]
  messages: ReportMessageDto[]
  generatedSqlQueries: GeneratedSqlQueryDto[]
}